# steamphony-api/openapi.yaml
openapi: 3.1.0
info:
  title: Steamphony API
  version: 1.0.0
  description: >-
    REST API for Steamphony marketing platform. All responses use the unified
    envelope structure.
servers:
  - url: https://api.steamphony.com
    description: Production
  - url: http://localhost:3001
    description: Local development
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
    AuthResponse:
      type: object
      properties:
        token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        user:
          type: object
          properties:
            id: { type: string }
            email: { type: string }
            role: { type: string }
      required: [token, user]
    LeadResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Contact form submitted successfully
        lead:
          type: object
          properties:
            id:
              type: string
              example: clx123abc
            name:
              type: string
            email:
              type: string
            createdAt:
              type: string
              format: date-time
      required: [success, message, lead]
    HealthResponse:
      type: object
      properties:
        status:
          type: string
          example: OK
        timestamp:
          type: string
          format: date-time
        environment:
          type: string
          example: development
paths:
  /api/health:
    get:
      summary: Health check
      responses:
        '200':
          description: Health status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
  /api/contact:
    post:
      summary: Submit contact form (lead capture)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                phone:
                  type: string
                businessType:
                  type: string
                  enum: [restaurant, beauty, retail, other]
                message:
                  type: string
                language:
                  type: string
                  example: ru
              required: [name, email, message]
      responses:
        '201':
          description: Lead created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeadResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/analytics/event:
    post:
      summary: Send analytics event (internal)
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                event:
                  type: string
                data:
                  type: object
                  additionalProperties: true
              required: [event]
      responses:
        '200':
          description: Event recorded
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
        '401':
          description: Invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /live:
    get:
      summary: Liveness probe
      responses:
        '200':
          description: Plain OK
          content:
            text/plain:
              schema:
                type: string
                example: OK
  /ready:
    get:
      summary: Readiness probe (checks DB)
      responses:
        '200':
          description: Ready
          content:
            text/plain:
              schema:
                type: string
                example: READY
        '503':
          description: Service unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /metrics:
    get:
      summary: Prometheus metrics
      responses:
        '200':
          description: Metrics in Prometheus exposition format
          content:
            text/plain:
              schema:
                type: string
  /api/auth/register:
    post:
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string }
                password: { type: string, format: password }
              required: [email, password]
      responses:
        '201':
          description: JWT token and user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict (user exists)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/login:
    post:
      summary: Login existing user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string }
                password: { type: string, format: password }
              required: [email, password]
      responses:
        '200':
          description: JWT token and user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/admin/ping:
    get:
      summary: Ping protected admin route
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Pong
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  message: { type: string, example: pong }
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse' 